Finished the Hundo Project
Following the Mini-Site Template parts 1-4 & Building the hundo videos. We built the design of the website, used CSS /Bootstrap to modify the look of it, and typed in the JavaScript for the app itself. We then added the code text from the site.js to the code page via prism.js so that the user could look over the essential code of the Hundo Application. Then I commented out the social media platforms I currently don't use, and plan on uploading a new logo and my current social media platforms later.